rhi: Add the more appropriate srb class name as an alias
We will not have time to roll out the proper name change
in 6.7, even though we are not bound by compatibility promises.

In the absence of a holistic migration that covers all Qt modules,
just add the new name with using. The proper solution would be the
reverse: the name changes everywhere and the old class name becomes
available via using; but for now we can live with the reverse,
which still allows new code (such as new examples, and future
application code) use the more descriptive name.

Change-Id: I18e791f4f93cb9b4fae8ec21aeb99b4bf93662f1
Reviewed-by: Andy Nichols <andy.nichols@qt.io>