JNI: remove the environment check in QJniObject again
This amends 944200b5a9705a7617f82cdaf5caf8932380aba4 and the follow-up
commit 84e70976f3d3df8ac921877d32c1dd884fd64267. We have too many use
cases in Qt where we construct a QJniObject from one thread, and access
it from others. Remove the overhead of checking and warning about the
mismatches.

The responsibility of guarding access to the underlying Java object, by
using MonitorEnter/Exit or synchronization blocks in Java, is with the
caller.

Change-Id: Iadbc9af0476009f2d43ecdd3d8f1335f31a04446
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>