CMake: Move QtAutoDetect.cmake functions into a separate Helpers file
And automatically include and install it in all qt builds.
We will use some of its functions for enforcing apple sdk
requirements when building Qt.

For the 6.6 branch there was a conflict that vcpkg integration is
always ON, but was changed to be OFF in 6.7.
We keep the 6.6 behavior, aka vcpkg integration will be ON.

Change-Id: I46383bc857430d35314dfa2ebef9eb342fb63560
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
Reviewed-by: Amir Masoud Abdol <amir.abdol@qt.io>
(cherry picked from commit ba9d45088f966430a814e7af36a8e7657d61e6db)